+++ src/tool/subcommands/snapshot_cmd.rs
@@ -6,17 +6,24 @@ use crate::blocks::Tipset;
 use crate::chain::index::ChainIndex;
 use crate::cli_shared::snapshot;
 use crate::daemon::bundle::load_actor_bundles;
+use crate::db::car::forest::DEFAULT_FOREST_CAR_FRAME_SIZE;
 use crate::db::car::ManyCar;
 use crate::ipld::{recurse_links_hash, CidHashSet};
 use crate::networks::{calibnet, mainnet, ChainConfig, NetworkChain};
 use crate::shim::machine::MultiEngine;
+use crate::utils::db::car_stream::CarStream;
 use crate::utils::proofs_api::paramfetch::ensure_params_downloaded;
 use anyhow::{bail, Result};
 use cid::Cid;
 use clap::Subcommand;
+use dialoguer::{theme::ColorfulTheme, Confirm};
+use futures::TryStreamExt;
 use fvm_ipld_blockstore::Blockstore;
+use indicatif::{ProgressBar, ProgressStyle};
 use std::path::PathBuf;
 use std::sync::Arc;
+use tokio::fs::File;
+use tokio::io::AsyncWriteExt;
 
 #[derive(Debug, Subcommand)]
 pub enum SnapshotCommands {
@@ -47,6 +54,26 @@ pub enum SnapshotCommands {
         /// Path to a snapshot CAR, which may be zstd compressed
         #[arg(required = true)]
         snapshot_files: Vec<PathBuf>,
+    },
+
+    /// Make this snapshot suitable for use as a compressed car-backed blockstore.
+    Compress {
+        /// Input CAR file, in `.car`, `.car.zst`, or `.forest.car.zst` format.
+        source: PathBuf,
+        /// Output file, will be in `.forest.car.zst` format.
+        ///
+        /// Will reuse the source name (with new extension) if pointed to a
+        /// directory.
+        #[arg(short, long, default_value = ".")]
+        output_path: PathBuf,
+        #[arg(long, default_value_t = 3)]
+        compression_level: u16,
+        /// End zstd frames after they exceed this length
+        #[arg(long, default_value_t = DEFAULT_FOREST_CAR_FRAME_SIZE)]
+        frame_size: usize,
+        /// Overwrite output file without prompting.
+        #[arg(long, default_value_t = false)]
+        force: bool,
     },
 }
 
@@ -79,6 +106,68 @@ impl SnapshotCommands {
                     check_stateroots,
                 )
                 .await
+            }
+            Self::Compress {
+                source,
+                output_path,
+                compression_level,
+                frame_size,
+                force,
+            } => {
+                // If input is 'snapshot.car.zst' and output is '.', set the
+                // destination to './snapshot.forest.car.zst'.
+                let destination = match output_path.is_dir() {
+                    true => {
+                        let mut destination = output_path;
+                        destination.push(source.clone());
+                        while let Some(ext) = destination.extension() {
+                            if !(ext == "zst" || ext == "car" || ext == "forest") {
+                                break;
+                            }
+                            destination.set_extension("");
+                        }
+                        destination.with_extension("forest.car.zst")
+                    }
+                    false => output_path.clone(),
+                };
+
+                if !force && destination.exists() {
+                    let have_permission = Confirm::with_theme(&ColorfulTheme::default())
+                        .with_prompt(format!(
+                            "{} will be overwritten. Continue?",
+                            destination.to_string_lossy()
+                        ))
+                        .default(false)
+                        .interact()
+                        // e.g not a tty (or some other error), so haven't got permission.
+                        .unwrap_or(false);
+                    if !have_permission {
+                        return Ok(());
+                    }
+                }
+
+                println!("Generating forest.car.zst file: {:?}", &destination);
+
+                let file = File::open(&source).await?;
+                let pb = ProgressBar::new(file.metadata().await?.len()).with_style(
+                    ProgressStyle::with_template("{bar} {percent}%, eta: {eta}")
+                        .expect("infallible"),
+                );
+                let file = tokio::io::BufReader::new(pb.wrap_async_read(file));
+
+                let mut block_stream = CarStream::new(file).await?;
+                let roots = std::mem::take(&mut block_stream.header.roots);
+
+                let mut dest = tokio::io::BufWriter::new(File::create(&destination).await?);
+
+                let frames = crate::db::car::forest::Encoder::compress_stream(
+                    frame_size,
+                    compression_level,
+                    block_stream.map_err(anyhow::Error::from),
+                );
+                crate::db::car::forest::Encoder::write(&mut dest, roots, frames).await?;
+                dest.flush().await?;
+                Ok(())
             }
         }
     }